pyproject.toml
[project]
name = "project-3-kafka-petproject-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2026.1.4",
    "confluent-kafka[json]>=2.13.0",
    "faker>=40.1.2",
    "python-dotenv==1.2.1",
    "requests>=2.32.5",
]
